Update mamba (#187)
* Add explicit casting to fix result dtype as value-based promotion was removed from NumPy v2

+++ xwrf/grid.py
@@ -72,8 +72,8 @@ def _wrf_grid_from_dataset(ds: xr.Dataset) -> Mapping[Hashable, pyproj.CRS | np.
         trf = pyproj.Transformer.from_crs(wgs84, crs, always_xy=True)
         e, n = trf.transform(cen_lon, cen_lat)
 
-    nx = ds.sizes['west_east']
-    ny = ds.sizes['south_north']
+    nx = np.float64(ds.sizes['west_east'])
+    ny = np.float64(ds.sizes['south_north'])
     x0 = -(nx - 1) / 2.0 * dx + e  # DL corner
     y0 = -(ny - 1) / 2.0 * dy + n  # DL corner
